Simplifying Bible class removing ABC and using Apocrypha as a instance attribute
